requirements.txt
# WhatsApp FAQ Bot Requirements

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# HTTP Requests
requests==2.31.0

# OpenAI API
openai==1.3.0

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.9.1
flake8==6.1.0
isort==5.12.0

# Optional: For production deployment
gunicorn==21.2.0
waitress==2.1.2